security configured with jwt, db changed, added exception handler, added interfaces for controllers, refactoring

+++ src/main/java/com/soundy/service/PlaylistService.java
@@ -1,5 +1,6 @@
 package com.soundy.service;
 
+import com.soundy.dto.exception.OwnerInvalidException;
 import com.soundy.dto.playlist.CreatePlaylistReq;
 import com.soundy.dto.playlist.GetPlayListResp;
 import com.soundy.dto.playlist.PlaylistUpdateReq;
@@ -13,7 +14,9 @@ import com.soundy.repository.TrackRepository;
 import lombok.AllArgsConstructor;
 import org.springframework.stereotype.Service;
 
+import java.security.Principal;
 import java.util.HashSet;
+import java.util.Objects;
 import java.util.Optional;
 import java.util.Set;
 
@@ -27,12 +30,13 @@ public class PlaylistService {
 
     AppUserRepository userRepository;
 
+
     public Optional<GetPlayListResp> findPlayListById(Integer id) {
         return playlistRepository.findById(id).map(SoundyMapper.INSTANCE::toPlaylistResp);
     }
 
-    public boolean createPlaylist(CreatePlaylistReq req) {
-        Optional<AppUser> owner = userRepository.findById(req.getOwnerId());
+    public boolean createPlaylist(CreatePlaylistReq req, Principal principal) {
+        Optional<AppUser> owner = userRepository.findByUsername(principal.getName());
         Playlist playlist = SoundyMapper.INSTANCE.toPlaylist(req);
         if (owner.isPresent()) {
             Set<Track> tracks = new HashSet<>(trackRepository.findAllById(req.getTracks()));
@@ -43,11 +47,16 @@ public class PlaylistService {
         return false;
     }
 
-    public boolean updatePlaylist(PlaylistUpdateReq req) {
+    public boolean updatePlaylist(PlaylistUpdateReq req, Principal principal) throws OwnerInvalidException {
+        Optional<AppUser> owner = userRepository.findByUsername(principal.getName());
         Optional<Playlist> opt = playlistRepository.findById(req.getId());
-        if (opt.isPresent()) {
-            Set<Track> tracks = new HashSet<>(trackRepository.findAllById(req.getTracks()));
+        if (opt.isPresent() && owner.isPresent()) {
             Playlist playlist = opt.get();
+            AppUser user = owner.get();
+            if (!Objects.equals(playlist.getOwner().getId(), user.getId())) {
+                throw new OwnerInvalidException();
+            }
+            Set<Track> tracks = new HashSet<>(trackRepository.findAllById(req.getTracks()));
             playlist.setTracks(tracks).setTitle(req.getName()).setDesc(req.getDesc());
             playlistRepository.save(playlist);
             return true;
@@ -55,8 +64,19 @@ public class PlaylistService {
         return false;
     }
 
-    public void deletePlaylistById(Integer id) {
+    public boolean deletePlaylistById(Integer id, Principal principal) throws OwnerInvalidException {
+        var pl = playlistRepository.findById(id);
+        var usr = userRepository.findByUsername(principal.getName());
+        if (usr.isEmpty() || pl.isEmpty()) {
+            return false;
+        }
+        var user = usr.get();
+        var playlist = pl.get();
+        if (!Objects.equals(playlist.getOwner().getId(), user.getId())) {
+            throw new OwnerInvalidException();
+        }
         playlistRepository.deleteById(id);
+        return true;
     }
 
 }
